add use-offer-defaults.ts
